Initially, the number of queens is zero We will need to import namedtuples and several manually set Errors that are to be ignored to the unittest file in order for further manipulation.
Secondly, we are making a unittest class that includes and contains multiple tests to testify and verify if certain codes within the main code are executing correctly. For instance, as we check for the original clear board, there should not be any things that is placed on the board, so, we testify whether a plot is clear by using assertEqual function.

The functions that we are testifying includes:
-Plot is empty
-Whether an object is successfully added to the board or not
-Whether an object is successfully removed or not
-Whether is an object that is cued to be removed has successfully executed ot not
-Testify is the queen is unsafe according to its locations with 2 different tests.
...

And Eventually executing the code using the __name__ == '__main__' function to see if the unittest itself and its corresponding main codes has worked out successfully or not.