Initially, the number of queens is zero We will need to import namedTuples and several manually set Errors that are to be ignored to the unittest file in order for further manipulation.
Secondly, we are making a unittest class that includes and contains multiple tests to testify and verify if certain codes within the main code are executing correctly. For instance, as we check for the original clear board, there should not be any things that is placed on the board, so, we testify whether a plot is clear by using assertEqual function.

The functions that we are testifying includes:
-Plot is empty
-Whether an object is successfully added to the board or not
-Whether an object is successfully removed or not
-Whether is an object that is cued to be removed has successfully executed ot not
-Testify is the queen is unsafe according to its locations with 2 different tests.
...

And Eventually executing the code using the __name__ == '__main__' function to see if the unittest itself and its corresponding main codes has worked out successfully or not.

For the queens.py (main code):
We are first to initialize the board with -.
secondly, we set the function to count the total number of queens on the chess board by summing up the existing object that has been placed by using summing and iterable functions that checks repeatedly for whether a space has been filled up.

Thirdly we count the existing queens that has been placed down using the same iteration method above.

Fourthly we are to use multiples of mutually included iteration method to find and record the positions of the queens that has been placed down.

Fifthly we testify if a place has been filled up by a queen or not.

Sixth step I did is to check for if any queens are unsafe by any chance by using abs and iteration functions to check its locations and testify if it is 'safe'

The seventh step is to check if any queens are added to a given position which will raise a customized error if yes by using iterations again.

The last step is to testify and check if there is an execution of removing a queen on a spot where a queen does not even exist.